김준영 : Bazooka 스피드 변경, State 업데이트

+++ TeamApiGame_Worms/GameEngineContents/GameTurn.cpp
@@ -87,7 +87,7 @@ void GameTurn::ChangeTurnPlayer()
 	TurnPlayer->SwitchIsTurnPlayer();
 
 	++StartValue;
-	int SIZE = Player::GetAllPlayer().size();
+	size_t SIZE = Player::GetAllPlayer().size();
 	if (StartValue >= SIZE)
 	{
 		StartValue = 0;
